requirement.txt
streamlit
pyngrok
python-docx
pdfplumber
joblib
numpy
pandas
scikit-learn
pdfminer.six
pypdfium2
matplotlib
seaborn
imbalanced-learn
xgboost
